#1 Added new config options
New config options for disabling pronouns, a message fade duration,
emote size, excludes for bot- and command messages

Also removing messages that are out of bounds(i.e. scrolled out of the
viewport) by default now.

Fixed a but that would display pronoun as "(false)" if not set.

The new skip_message function is also preparation for blocklisting
or simple word ban lists.