Removing OBS copy link and background colour when applicable now
The OBS copy link won't work with local file:// URLs(it sets it to "Local file"
and disables all URL parameters) so we don't show it on local files

Also removing the user name background colour selector for when bubbles
aren't enabled, since there the option isn't used.

This also includes a bugfix for the recent changes in the colour
overrides when not using bubbles. This previously did not work as
intended because the naming of these variable is
suboptimal and could be improved upon in a refactor.

Also added some code documentation.